Move StaticSecurityHeaders to util
