+++ pkg/util/httputil/static_security_headers.go
@@ -1,10 +1,10 @@
-package web
+package httputil
 
-import "net/http"
+import (
+	"net/http"
+)
 
-type StaticSecurityHeadersMiddleware struct{}
-
-func (StaticSecurityHeadersMiddleware) Handle(next http.Handler) http.Handler {
+func StaticSecurityHeaders(next http.Handler) http.Handler {
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		w.Header().Set("X-Content-Type-Options", "nosniff")
 		next.ServeHTTP(w, r)
